+++ src/main/java/com/barco/common/CommonApplication.java
@@ -2,8 +2,6 @@ package com.barco.common;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import javax.annotation.PostConstruct;
-import java.util.TimeZone;
 
 /**
  * @author Nabeel Ahmed
@@ -13,11 +11,6 @@ public class CommonApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(CommonApplication.class, args);
-	}
-
-	@PostConstruct
-	public void started() {
-		TimeZone.setDefault(TimeZone.getTimeZone("America/New_York"));
 	}
 
 }
