+++ src/main/java/com/barco/common/manager/ftp/FtpFileExchange.java
@@ -7,7 +7,6 @@ import org.apache.commons.net.ftp.FTPReply;
 import org.apache.commons.net.ftp.FTPSClient;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.springframework.stereotype.Component;
 import javax.net.ssl.*;
 import java.io.IOException;
 import java.io.InputStream;
@@ -17,8 +16,6 @@ import java.lang.reflect.Method;
 import java.net.Socket;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
-import java.security.Provider;
-import java.security.Security;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.util.Locale;
@@ -26,7 +23,6 @@ import java.util.Locale;
 /**
  * @author Nabeel Ahmed
  */
-@Component
 public class FtpFileExchange {
 
     private Logger logger = LogManager.getLogger(FtpFileExchange.class);
