+++ src/main/resources/application-dev.properties
@@ -1,9 +1,9 @@
 #===============================
 # ==== JWT Token detail ===========
 # ===============================
-app.jwtSecret= bezKoderSecretKey
-app.jwtExpirationMs= 86400000
-app.jwtRefreshExpirationMs= 8640000
+app.jwtSecret=bezKoderSecretKey
+app.jwtExpirationMs=86400000
+app.jwtRefreshExpirationMs=8640000
 # ===============================
 # = Shareddle TaskExecutor
 # ===============================
