+++ src/main/resources/application-prod.properties
@@ -1,19 +1,19 @@
 #===============================
 # ==== JWT Token detail ===========
 # ===============================
-app.jwtSecret= bezKoderSecretKey
-app.jwtExpirationMs= 86400000
-app.jwtRefreshExpirationMs= 8640000
+app.jwtSecret=${JWT_SECRET}
+app.jwtExpirationMs=${JWT_EXPIRATION_MS}
+app.jwtRefreshExpirationMs=${JWT_REFRESH_EXPIRATION_MS}
 # ===============================
 # = Shareddle TaskExecutor
 # ===============================
-asyncetaskexecutor.minThreads=20
-asyncetaskexecutor.maxThreads=50
+asyncetaskexecutor.minThreads=${MIN_THREADS}
+asyncetaskexecutor.maxThreads=${MAX_THREADS}
 asyncetaskexecutor.idleThreadLife=1
 #===============================
 # ==== Email detail ===========
 # ===============================
-spring.mail.host=sandbox.smtp.mailtrap.io
+spring.mail.host=${MAIL_HOST}
 spring.mail.port=25
-spring.mail.username=488af377a62331
-spring.mail.password=827886bc8aaf08
+spring.mail.username=${MAIL_USERNAME}
+spring.mail.password=${MAIL_PASSWORD}
